pyproject.toml
[tool.poetry]
name = "ai-hedge-fund"
version = "0.1.0"
description = "AI Hedge Fund Simulator"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.0"
pandas = "^2.1.0"
pandas-datareader = "^0.10.0"
yfinance = "^0.2.35"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
plotly = "^5.18.0"
streamlit = "^1.27.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.0"
openai = "^2.0.0"
langchain = "^0.3.28"
langchain-community = "^0.3.26"
langsmith = "^0.3.45"
langchain-core = "^0.3.73"
#langchain-openai ="^1.1.14"
#langchain-groq="^1.1.2"
chromadb = "^0.4.22"
tiktoken = "^0.7.0"
groq = "^0.4.0"
alpaca-py = "^0.43.2"
websockets = "^12.0"
backtrader = "^1.9.78"
pyfolio = "^0.9.2"
empyrical = "^0.5.5"
quantstats = "^0.0.62"
ta = "^0.10.2"
schedule = "^1.2.0"
tzdata = "^2023.3"
python-telegram-bot = "^22.7"


[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"